requirements.txt
Flask
streamlit
langchain
python-dotenv
transformers
